Moved libagt, libdraldb and dbtest from asimcore
The moved sources are the current HEAD version in asimcore
They are a partial port to qt4. Ongoing qt4 porting effort should
happen now in this branch (qt4porting branch in the dreams repository)

The configure.in and the autotool output files have been changed in order to look for qt4 instead of qt3

At least 1Dreams, Dreams and the libagt library do not compile yet with qt4.
Such sources are behind the @TODO@ variable in the corresponding Makefile.am files (so they do not get compiled)
Once the porting effort is done, they should be moved to before the @TODO@ variable